requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
azure-core==1.24.2
azure-storage-blob==12.12.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blis==0.7.7
cachetools==5.1.0
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==37.0.2
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.5
django-countries==7.3.2
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.3.0/en_core_web_md-3.3.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
fake-factory==9999.9.9
Faker==13.12.1
fastjsonschema==2.15.3
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
msrest==0.7.1
murmurhash==1.0.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pickleshare==0.7.5
Pillow==9.1.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyarrow==8.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
semver==2.13.0
Send2Trash==1.8.0
shell==1.0.1
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-streamlit==1.0.3
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.2.0
streamlit==1.9.0
terminado==0.15.0
thinc==8.0.16
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uuid==1.30
validators==0.19.0
waitress==2.1.2
wasabi==0.9.1
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
widgetsnbextension==3.6.0
zipp==3.8.0
